requirements.txt
# GridSense AI Prototype Requirements
# Python 3.10+ required

# Core data & ML
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.18.0
altair>=5.2.0

# Time Series Forecasting (Chronos)
chronos-forecasting>=2.2.0
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# Computer Vision / Satellite
opencv-python>=4.8.0
Pillow>=10.0.0

# Weather API
openmeteo-requests>=1.2.0
requests-cache>=1.1.0
retry-requests>=2.0.0

# API & Dashboard
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6

# Physics / Math
scipy>=1.11.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Config
python-dotenv>=1.0.0

# Optional: Satpy for MOSDAC HDF5 (can be tricky to install; fallback to h5py)
# satpy>=0.60.0
h5py>=3.10.0
xarray>=2023.10.0
